requirements.txt
flask>=3.0,<4
supabase>=2.0,<3
PyJWT[crypto]>=2.8,<3
requests>=2.31,<3
gunicorn>=21.2,<23
